Cleanup: Move /api/queries/queryList.ts to /api/queries.ts
