+++ mathesar_ui/src/components/QueryName.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { QueryInstance } from '@mathesar/api/queries/queryList';
+  import type { QueryInstance } from '@mathesar/api/queries';
   import { iconQuery } from '@mathesar/icons';
   import NameWithIcon from './NameWithIcon.svelte';
