+++ mathesar_ui/src/systems/query-builder/QueryModel.ts
@@ -1,7 +1,7 @@
 import type {
   QueryInstanceInitialColumn,
   QueryInstanceTransformation,
-} from '@mathesar/api/queries/queryList';
+} from '@mathesar/api/queries';
 import type { UnsavedQueryInstance } from '@mathesar/stores/queries';
 import QueryFilterTransformationModel from './QueryFilterTransformationModel';
 import QuerySummarizationTransformationModel from './QuerySummarizationTransformationModel';
