+++ mathesar_ui/src/systems/query-builder/QuerySummarizationTransformationModel.ts
@@ -1,4 +1,4 @@
-import type { QueryInstanceSummarizationTransformation } from '@mathesar/api/queries/queryList';
+import type { QueryInstanceSummarizationTransformation } from '@mathesar/api/queries';
 import { ImmutableMap } from '@mathesar/component-library';
 
 export interface QuerySummarizationAggregationEntry {
